feat: implement backend orchestration for AI website generator
This commit introduces a new backend orchestration layer to manage the multi-step API workflow for the AI website generator.

The key changes are:
- A new API route at `/api/ai/orchestration/route.ts` has been created to handle the end-to-end orchestration logic.
- The existing API routes (`chat`, `plan`, `registry`, `select-components`, `get-code`, `edit-code`, `generate-code`, `color-palette`) have been refactored to export their core logic, which is now called directly by the orchestrator.
- The `AIChat.tsx` component has been updated to call the new orchestration endpoint and handle the streamed JSON response.
- The client-side orchestration logic has been removed from `AIChat.tsx`, resulting in a cleaner and more maintainable frontend component.

This change addresses the user's request to move the orchestration logic to the backend, providing a more robust, secure, and efficient solution.
