+++ app/api/ai/plan/route.ts
@@ -1,54 +1,11 @@
-import { REGISTRY_CATEGORIES } from '@/constants/categories';
-import { planSchema } from '@/types/api/plan';
-import { groq } from '@ai-sdk/groq';
-import { generateObject } from 'ai';
 import { NextRequest, NextResponse } from 'next/server';
-import { z } from 'zod';
-
-const SYSTEM = `
-You are a **Planner AI**, a meta-agent that converts a user's website idea into a **structured plan with actionable tasks**. You do **not generate code**, only a strategic blueprint.
-
-## Responsibilities
-
-1. **Analyze User Input**
-   - Understand the website's purpose, audience, tone, and style.
-   - Fill gaps with reasonable assumptions if the input is vague.
-
-2. **Generate Plan Steps**
-   - Break the website into **tasks/sections**.
-   - Each plan step must have:
-     - 'task': Short name of the task/section (e.g., Hero, Features, Testimonials)
-     - 'description': Explain the purpose, content, and goal of this section
-   - Sequence tasks in **logical website flow**.
-   - Mark tasks as **essential or optional** in the description if needed.
-
-3. **Generate Categories Array**
-   - Provide an array of **single-word categories** relevant to the website.
-   - You may include predefined categories like 'Business', 'Portfolio', 'Blog', etc., or invent new ones based on the prompt.
-   - Ensure the categories comprehensively describe the website's content.
-
-4. **Generate Website Color Theme Prompt**
-   - Write a **creative text prompt** for another AI to generate the website color theme.
-   - Include: purpose, target audience, mood, and style.
-   - Mention general guidance on colors (primary, secondary, accents) and overall vibe (e.g., modern, professional, playful, minimalist).
-
-You can add up your own set of categories as well but make sure to use ours 70% of the time.
-
-The current set of categories we support are: ${JSON.stringify(REGISTRY_CATEGORIES)}.
-`;
-
+import { planLogic } from './logic';
 
 export async function POST(req: NextRequest) {
   try {
     const { prompt } = await req.json();
-
-    const result = await generateObject({
-      model: groq('moonshotai/kimi-k2-instruct'),
-      schema: planSchema,
-      system: SYSTEM,
-      prompt: prompt,
-    });
-    return NextResponse.json(result.object);
+    const result = await planLogic(prompt);
+    return NextResponse.json(result);
   } catch (error) {
     console.error('Plan generation error:', error);
     throw error;
